✨ F08: ADMIN > AddProduct > Create new product with form informations when submited
