+++ src/context/OrderContext.jsx
@@ -3,7 +3,7 @@ import { createContext } from "react";
 export default createContext({
 	isModeAdmin: false,
 	setIsModeAdmin: () => {},
-	menu: {},
+	menu: [],
 	setMenu: () => {},
 	isCollapsed: false,
 	setIsCollapsed: () => {},
